requirements.txt
prometheus-client
requests
pyjq
inflection
pyyaml